requirements.txt
cloudinary==1.15.0
dj-database-url==0.5.0
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
django_heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.7
PyJWT==1.7.1
python-dotenv==0.10.1
pytz==2018.9
django-crontab==0.7.1
pillow==6.0.0
whitenoise==4.1.2
